Babel, ESLint, BisectRun, Timing, Verbose, CPUs
This project is now built with Babel to enable the use of ES6. I've
also introduced ESLint for consistency.

The BisectRun now encapsulates all data that is specific to a given
execution of the tool.

Added a `--verbose` option for debugging and gathering more information
about how the tool is viewing your tests.

Parallelization is now pegged to the number of CPUs on the machine.

Added timing information to the output.