Fix improper queueing with in-flight child processes
The check for whether to queue a request or service it immediately
checked the queue length, but not the in-flight count. As a result
the ExecQueue would exceed its specified concurrency level.

The check in enqueue now compares against the sum of in-flight and
queue length to ensure proper concurrency is maintained.